Event bus

+++ Interfaces/IPatient.cs
@@ -8,12 +8,12 @@ namespace Interfaces
 {
     public interface IPatient
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; }
 
         public DateTime Birthday { get; set; }
 
-        public long MedicalHistoryNumber { get; set; }
+        public int MedicalHistoryNumber { get; set; }
     }
 }
